Switch back to rspec
A long time ago, the project started with rspec and moved to test-unit.
Now we're back to rspec, because if you can't beat em, join em.

Notably, this commit also upgrades the project to Ruby 2.0. No breaking
changes were observed, as it's mostly backwards compatible. The eventual
shift to Rails 4 will come with more code.