Switch to asynchronous mailers
Synchronous mailers are slow and make me sad. Async mailers are a
low-hanging fruit for optimizing performance of controller actions. This
is accomplished by adding Sidekiq (which depends on redis) and using
its ActiveMailer extensions to send mails through its queueing backend.

The signature of mailer methods is redefined to generally accept IDs
over ActiveRecord objects (which have to be serialized to the queue
before the mail is sent)