Initial payments infrastructure
WIP

The release of Square Cash provides a unique service that enables us to
integrate paying off debts entirely within the platform without actually
processing any transactions. The extent of integration is basically
adding a mailto link with fields pre-formatted; however, a compelling
design and platform integration drive the feature.

A payments button is provided on the dashboard to send a Square Cash
email; this button is also instrumented to send data to the Payback
backend so we can track and display payment data.

As of now, we have no way to confirm whether the user has properly
registered or payment was successful - it's a 'fire and forget' system.
This should work for the time being, although it would be possible to
add a contention / reversal mechanism if need be.

The feature announcement page is a work in progress.

Other additions include:

- The nav bar is refactored to use stack dropdowns on mobile
  due to the addition of the Payments tab.

- The pages directory is removed from app/assets/stylesheets. That was
  dumb.

- The static layout is removed, and the logic is compressed into
  application.html.erb.

- Tipsy tooltips are refactored - instead of manually calling .tipsy()
  on each selector, a declarative data-tipsy attribute is now used.