First implementation of payments for net totals
The very first implementation of payments was centered around the
expense. However, it'd be a convenience to be able to pay a net total
owed at once in a single payment. This adds a 'pay all button' to the
user debts page, which will automatically configure a payment with the
proper amount.

The difficulty is in tracking the mapping back to the original expenses.
If we create a single payment record with the amount equal to the total
amount owed, we have no mapping back to the expenses that was for, and
have no way to display that a particular expense has already been paid
for.

An alternate approach would be to create a payment record for each
expense paid for. However, this could be confusing to the user as the
payment was a single transfer representing a contiguous sum, and still
doesn't solve the problem of figuring how whether or not a user has paid
back another user in full.

This commit implements the first approach (a single total payment) - the
mapping item is left as a TODO.