First attempt at notifications refactor
The new payments notification type can correspond to multiple expenses
("Jill sent $99 for 3 expenses"). This requires a refactor of the
notifications infrastructure as previously a notification was tied to a
single expense.

The mark off button on the payments notification is still a work in
progress - it will work exactly the same as the mark off request type,
but needs to target multiple expenses.