Finish payment notification JS interactions
This commit implements a similar mark off button / slide up
functionality for the payments notification button, which can
potentially be associated with multiple payments. Included is an option
to toggle a list of expense titles that a payment is associated with.

One final TODO is how to deal with outstanding credits from a user when
paying off a net debt. It makes the most sense to automatically mark
them off, but that needs to be clear to the user. Things could
potentially be confusing while waiting for the user to acknowledge the
payment, but the envelope icon should address this.